- refactor, rearrange codes and structures
